refactor: deconstruct WebLlmSettingsSubComponents into modular files and adopt Button atom
