+++ memory-bank/userJourneys.json
@@ -610,7 +610,14 @@
       "viewComponents": [
         "src/components/molecules/WebGpuWarning.tsx",
         "src/components/organisms/WebLlmSettingsSection.tsx",
-        "src/components/organisms/WebLlmSettingsSubComponents.tsx",
+        "src/components/organisms/WebLlmSettings/WebLlmStatusRow.tsx",
+        "src/components/organisms/WebLlmSettings/WebLlmProgress.tsx",
+        "src/components/organisms/WebLlmSettings/WebLlmRetryingInfo.tsx",
+        "src/components/organisms/WebLlmSettings/WebLlmError.tsx",
+        "src/components/organisms/WebLlmSettings/WebLlmQuotaWarning.tsx",
+        "src/components/organisms/WebLlmSettings/WebLlmBothUnsupportedWarning.tsx",
+        "src/components/organisms/WebLlmSettings/WebLlmActionButtons.tsx",
+        "src/components/organisms/WebLlmSettings/WebLlmHeader.tsx",
         "src/components/molecules/AiDownloadStatus.tsx",
         "src/components/molecules/GlobalDownloadIndicator.tsx",
         "src/lib/ai/opfs-cache-manager.ts",
